refactor(app): clean
