+++ app/src/services/utils.js
@@ -1,3 +1,5 @@
+import { slugify } from "@/utils/string";
+
 export const timeSince = (date) => {
   let seconds = Math.floor((new Date() - date) / 1000);
   let interval = seconds / 31536000;
@@ -51,21 +53,7 @@ const downloadFile = (filename, csv) => {
   }
 };
 
-export const slugify = (string) => {
-  const a = "àáäâãåăæçèéëêǵḧìíïîḿńǹñòóöôœṕŕßśșțùúüûǘẃẍÿź·/_,:;";
-  const b = "aaaaaaaaceeeeghiiiimnnnoooooprssstuuuuuwxyz------";
-  const p = new RegExp(a.split("").join("|"), "g");
-  return string
-    .toString()
-    .toLowerCase()
-    .replace(/\s+/g, "-") // Replace spaces with -
-    .replace(p, (c) => b.charAt(a.indexOf(c))) // Replace special characters
-    .replace(/&/g, "-and-") // Replace & with 'and'
-    .replace(/[^\w-]+/g, "") // Remove all non-word characters
-    .replace(/--+/g, "-") // Replace multiple - with single -
-    .replace(/^-+/, "") // Trim - from start of text
-    .replace(/-+$/, ""); // Trim - from end of text
-};
+export { slugify };
 
 export const isValidUrl = (url) => {
   try {
