+++ app/src/utils/string.js
@@ -1,0 +1,15 @@
+export const slugify = (string) => {
+  const a = "àáäâãåăæçèéëêǵḧìíïîḿńǹñòóöôœṕŕßśșțùúüûǘẃẍÿź·/_,:;";
+  const b = "aaaaaaaaceeeeghiiiimnnnoooooprssstuuuuuwxyz------";
+  const p = new RegExp(a.split("").join("|"), "g");
+  return string
+    .toString()
+    .toLowerCase()
+    .replace(/\s+/g, "-")
+    .replace(p, (c) => b.charAt(a.indexOf(c)))
+    .replace(/&/g, "-and-")
+    .replace(/[^\w-]+/g, "")
+    .replace(/--+/g, "-")
+    .replace(/^-+/, "")
+    .replace(/-+$/, "");
+};
